server/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
influxdb-client>=1.44.0
python-dotenv>=1.0.0
